Update Q.2
Do the following using PANDAS Series:
Create a series with 5 elements. Display the series sorted on index and also sorted on values seperately Create a series with N elements with some duplicate values. Find the minimum and maximum ranks assigned to the values using ‘first’ and ‘max’ methods
Display the index value of the minimum and maximum element of a Series